Gh file to store binary data
This file is not a document of normal use. It's used exclusively to
store binary data of bricks. In any case it's advisable to open this
file from the editor of Grasshopper, all content is serialized in a
custom way, so if you save (overwrite) the document, will lose all data
and may not be read by the brick box.